Make wildlife extractor a little more robust
Error-log all remaining return-none cases, even those without useful
exception data.

Always return a tuple, so you can always unpack it. Instead of the whole
return being None, now only image will be.

As a result, the worker no longer crashes. Give the inky a text-plain
error to show, rather than a 500 for it to go into error mode for.

Feign a normal desktop Windows Firefox User-Agent. This doesn't really
seem to be the problem for why sometimes the regex to find the page for
the week gallery isn't matching, though. It's intermittent enough that
it might be some nondeterminism in how they're generating the HTML, but
it's not actually terribly soupy. Dunno. Will probably need to log the
full response to debug that (ick).