Be more patient when double-clearing PRI
PRI decode, since it's streaming, must hold the socket during the
pre-clear if _DOUBLE_UPDATE_CLEAR is enable. Make the socket timeout 40s
more patient in this case, about the time a display update takes.

I *believe* updating the timeout here works (the documentation is quite
ambiguous on the semantics...it might even be a full TIMEOUT + 40 on
the next read call). This is rare to reproduce (I should probably
generate a pathologically slow PRI to decode as a test case, but, well,
I haven't).