Disable double-clear in example config
It turns out this is some kind of problem with the constant powered-up
state when on USB. When powering via the battery connector, "burn in"
never seems to build up. And talking to a friend who turns out has a PhD
in e-ink displays, the weird pre-imaging is indeed probably correct.